Move system functionality into separate module.
