+++ src/cache_cargo_home.rs
@@ -173,7 +173,7 @@ async fn build_cached_folder_info(cache_type: CacheType) -> Result<CachedFolderI
 
 fn build_cache_entry(cache_type: CacheType, key: &HashValue, path: &Path) -> CacheEntry {
     use crate::cache_key_builder::CacheKeyBuilder;
-    use crate::date;
+    use crate::system::date;
 
     let name = cache_type.friendly_name();
     let mut key_builder = CacheKeyBuilder::new(&name);
@@ -261,7 +261,7 @@ pub async fn restore_cargo_cache() -> Result<(), Error> {
 }
 
 pub async fn save_cargo_cache() -> Result<(), Error> {
-    use crate::date;
+    use crate::system::date;
     use humantime::format_duration;
     use wasm_bindgen::JsError;
