+++ src/lib.rs
@@ -16,7 +16,6 @@ mod cargo_hook;
 mod cargo_install_hook;
 mod cargo_lock_hashing;
 mod cross;
-mod date;
 mod dir_tree;
 mod error;
 mod fingerprinting;
@@ -24,11 +23,10 @@ mod node;
 mod nonce;
 mod noop_stream;
 mod package_manifest;
-mod rng;
+mod system;
 mod run;
 mod rustup;
 mod safe_encoding;
-mod sleep;
 mod toolchain;
 mod utils;
